Added fullscreen button
